refactor(client): Extract IntoUrl trait to module
