+++ src/client/mod.rs
@@ -70,6 +70,7 @@ use header::{ContentLength, Location};
 use method::Method;
 use net::{NetworkConnector, NetworkStream, Fresh};
 use {Url};
+use into_url::{IntoUrl};
 use Error;
 
 pub use self::pool::Pool;
@@ -393,30 +394,6 @@ impl<'a, R: Read> From<&'a mut R> for Body<'a> {
     #[inline]
     fn from(r: &'a mut R) -> Body<'a> {
         Body::ChunkedBody(r)
-    }
-}
-
-/// A helper trait to convert common objects into a Url.
-pub trait IntoUrl {
-    /// Consumes the object, trying to return a Url.
-    fn into_url(self) -> Result<Url, UrlError>;
-}
-
-impl IntoUrl for Url {
-    fn into_url(self) -> Result<Url, UrlError> {
-        Ok(self)
-    }
-}
-
-impl<'a> IntoUrl for &'a str {
-    fn into_url(self) -> Result<Url, UrlError> {
-        Url::parse(self)
-    }
-}
-
-impl<'a> IntoUrl for &'a String {
-    fn into_url(self) -> Result<Url, UrlError> {
-        Url::parse(self)
     }
 }
