+++ src/into_url.rs
@@ -1,0 +1,26 @@
+use url::Url;
+use url::ParseError as UrlError;
+
+/// A helper trait to convert common objects into a Url.
+pub trait IntoUrl {
+    /// Consumes the object, trying to return a Url.
+    fn into_url(self) -> Result<Url, UrlError>;
+}
+
+impl IntoUrl for Url {
+    fn into_url(self) -> Result<Url, UrlError> {
+        Ok(self)
+    }
+}
+
+impl<'a> IntoUrl for &'a str {
+    fn into_url(self) -> Result<Url, UrlError> {
+        Url::parse(self)
+    }
+}
+
+impl<'a> IntoUrl for &'a String {
+    fn into_url(self) -> Result<Url, UrlError> {
+        Url::parse(self)
+    }
+}
